requirements.txt
streamlit
nltk
scikit-learn
pandas
tensorflow
tensorflow-hub
Pillow
